程序集引用问题

+++ Mk.DemoC/src/Mk.DemoC.Application/ElastcSearchAppService/ElastcSearchTestAppService.cs
@@ -2,10 +2,9 @@
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
 using Mk.DemoC.Dto.ElastcSearchs;
-using Mk.DemoC.SearchDocumentMgr;
 using Mk.DemoC.SearchDocumentMgr.Documents;
 using Nest;
-using Swashbuckle.Swagger.Annotations;
+using Swashbuckle.AspNetCore.Annotations;
 using System.Collections.Generic;
 using System.Linq;
 using System.Net;
@@ -97,7 +96,7 @@ namespace Mk.DemoC.ElastcSearchAppService
         /// </summary>
         /// <returns></returns>
         [HttpPost("doc/init")]
-        [SwaggerResponse(HttpStatusCode.OK, Type = typeof(ServiceResult))]
+        [SwaggerResponse((int)HttpStatusCode.OK, Type = typeof(ServiceResult))]
         public async Task<ServiceResult<List<ProductSpuDocumentDto>>> InitDocument()
         {
             ServiceResult<List<ProductSpuDocumentDto>> ret = new ServiceResult<List<ProductSpuDocumentDto>>(IdProvider.Get());
@@ -192,7 +191,7 @@ namespace Mk.DemoC.ElastcSearchAppService
         }
 
         [HttpPost("doc/get")]
-        [SwaggerResponse(HttpStatusCode.OK, Type = typeof(ServiceResult<ProductSpuDocumentDto>))]
+        [SwaggerResponse((int)HttpStatusCode.OK, Type = typeof(ServiceResult<ProductSpuDocumentDto>))]
         public async Task<ServiceResult<List<ProductSpuDocumentDto>>> GetAsync()
         {
             ServiceResult<List<ProductSpuDocumentDto>> ret = new ServiceResult<List<ProductSpuDocumentDto>>(IdProvider.Get());
@@ -230,7 +229,7 @@ namespace Mk.DemoC.ElastcSearchAppService
         }
 
         [HttpPost("doc/update")]
-        [SwaggerResponse(HttpStatusCode.OK, Type = typeof(ServiceResult))]
+        [SwaggerResponse((int)HttpStatusCode.OK, Type = typeof(ServiceResult))]
         public async Task<ServiceResult> UpdateDocumentAsync()
         {
             string id = "39f792fdc909816aeda1bbb97faa18a5";
